requirements.txt
opencv-python
pandas
easydict
tqdm
matplotlib
numpy
albumentations[imgaug]
pyyaml
scikit-learn
wandb
timm
umap-learn
plotly
pytorch_warmup
imageio
openpyxl
imagededup
imagesize
open_clip_torch
onnx
onnxruntime
onnxruntime-gpu
wget
gdown
hydra-core